+++ state-machines/binary/src/binary_extension_collector.rs
@@ -62,7 +62,7 @@ impl BusDevice<u64> for BinaryExtensionCollector {
 
         self.inputs.push(BinaryInput::from(&data));
 
-        self.inputs.len() < self.num_operations as usize
+        self.inputs.len() < self.num_operations
     }
 
     /// Returns the bus IDs associated with this instance.
